+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [10.6.0] (melonJS 2) - _2022-xx-xx_
 
+### Changed
+- Renderable : refactor of Text and BitmapText to isolate metrics related features/functions and make them generic between the 2 classes
+
 ## [10.5.2] (melonJS 2) - _2022-03-17_
 
 ### Fixed